move TemplateApp files under src/, update Framework tsconfig to composite true so TS treats it as buildable
