+++ apps/TemplateApp/tsconfig.json
@@ -11,12 +11,8 @@
     "isolatedModules": true,
     "noEmit": true
   },
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "node_modules"
-  ],
+  "include": ["src"],
+  "exclude": ["dist", "node_modules"],
   "references": [
     {
       "path": "../../packages/Framework"
