+++ packages/Framework/tsconfig.json
@@ -3,13 +3,15 @@
     "target": "ES2022",
     "module": "CommonJS",
     "jsx": "react-jsx",
+    "composite": true,
     "declaration": true,
+    "declarationMap": true,
     "outDir": "dist",
     "rootDir": "src",
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
   },
   "include": ["src"],
   "exclude": ["dist", "node_modules"]
